# By Jose Sebastian Battig
# Via Jose Sebastian Battig
* 'fix_64_msvc_warnings' of https://github.com/Convey-Compliance/libbson:
  B-0 removed 64 bits warnings compiling in Visual Studio

Conflicts:
	src/bson/bson.c

+++ src/bson/bson.c
@@ -2256,7 +2256,7 @@ bson_compare (const bson_t *bson,
    ret = memcmp (data1, data2, BSON_MIN (len1, len2));
 
    if (ret == 0) {
-      ret = len1 - len2;
+      ret = (int64_t) (len1 - len2);
    }
 
    return (ret < 0) ? -1 : (ret > 0);
